- Prevent Sleep bug by rounding value to 10 to prevent http://www.javamex.com/tutorials/threads/sleep_issues.shtml
- Improve by switching from Execute to Timer

git-svn-id: http://goldengate.svn.sourceforge.net/svnroot/goldengate/GoldenGateFtp/trunk@1228 7864f6d1-e753-4b29-85b9-74ae93846b71

+++ src/main/java/goldengate/ftp/core/config/FtpConfiguration.java
@@ -128,7 +128,7 @@ public abstract class FtpConfiguration {
     /**
      * Nb of milliseconds after connection is in timeout
      */
-    public int TIMEOUTCON = 30000;
+    public long TIMEOUTCON = 30000;
 
     /**
      * Size by default of block size for receive/sending files. Should be a
